+++ HISTORY.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Added search route
 - Added index route
 - Added config file configuration
 - Added CLI